+++ noodles-bam/src/reader.rs
@@ -10,11 +10,12 @@ pub use self::{query::Query, records::Records, unmapped_records::UnmappedRecords
 use std::{
     ffi::CStr,
     io::{self, Read, Seek},
+    ops::Bound,
 };
 
 use byteorder::{LittleEndian, ReadBytesExt};
 use noodles_bgzf as bgzf;
-use noodles_core::{region::Interval, Region};
+use noodles_core::{region::Interval, Position, Region};
 use noodles_csi::{binning_index::ReferenceSequenceExt, BinningIndex};
 use noodles_fasta as fasta;
 use noodles_sam::{
@@ -321,7 +322,12 @@ where
     {
         let (reference_sequence_id, interval) = resolve_region(reference_sequences, region)?;
         let chunks = index.query(reference_sequence_id, interval)?;
-        Ok(Query::new(self, chunks, reference_sequence_id, interval))
+        Ok(Query::new(
+            self,
+            chunks,
+            reference_sequence_id,
+            region.interval(),
+        ))
     }
 
     /// Returns an iterator of unmapped records after querying for the unmapped region.
@@ -480,10 +486,19 @@ pub(crate) fn bytes_with_nul_to_string(buf: &[u8]) -> io::Result<String> {
         })
 }
 
+#[allow(clippy::type_complexity)]
 pub(crate) fn resolve_region(
     reference_sequences: &ReferenceSequences,
     region: &Region,
-) -> io::Result<(usize, Interval)> {
+) -> io::Result<(usize, (Bound<Position>, Bound<Position>))> {
+    fn cast_bound_i32_to_bound_position(bound: Bound<i32>) -> Bound<Position> {
+        match bound {
+            Bound::Included(n) => Bound::Included(Position::try_from(n as usize).unwrap()),
+            Bound::Excluded(n) => Bound::Excluded(Position::try_from(n as usize).unwrap()),
+            Bound::Unbounded => Bound::Unbounded,
+        }
+    }
+
     let i = reference_sequences
         .get_index_of(region.name())
         .ok_or_else(|| {
@@ -496,7 +511,12 @@ pub(crate) fn resolve_region(
             )
         })?;
 
-    Ok((i, region.interval()))
+    let interval = (
+        cast_bound_i32_to_bound_position(region.start()),
+        cast_bound_i32_to_bound_position(region.end()),
+    );
+
+    Ok((i, interval))
 }
 
 #[cfg(test)]